Decompile bmtarget

+++ include/bmphase.h
@@ -4,7 +4,7 @@
 int GetPhaseAbleUnitCount(int faction);
 int CountUnitsInState(int faction, int state);
 s8 AreUnitsAllied(int left, int right);
-int IsSameAllegiance(int left, int right);
+s8 IsSameAllegiance(int left, int right);
 int GetCurrentPhase(void);
 int GetNonActiveFaction(void);
